# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chama by Aventum - Digital Rotational Savings Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List organizations
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationBody"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{orgId}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization details
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationDetails"

  # GROUPS
...
  /contributions:
    get:
      operationId: listContributions
      tags: [contributions]
      summary: List contributions for current user
      parameters:
        - in: query
          name: groupId
          schema:
            type: integer
        - in: query
          name: cycleId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [paid, pending, failed]
      responses:
        "200":
          description: List of contributions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contribution"
  /contributions/pay:
    post:
      operationId: payContribution
      tags: [contributions]
      summary: Submit a contribution payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayContributionBody"
      responses:
        "201":
          description: Contribution recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contribution"
        "400":
          description: Already paid or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contributions/all:
    get:
      operationId: listAllContributions
      tags: [contributions]
      summary: List all contributions (admin)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All contributions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContributionListResponse"

  # PAYOUTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [member, group_admin, org_admin, super_admin, ceo, cto_admin, it_support, finance, marketing, relationship_manager]
        organizationId:
          type: integer
          nullable: true
        departmentId:
          type: integer
          nullable: true
        isFinanceAdmin:
          type: boolean
          nullable: true
        twoFactorEnabled:
          type: boolean
        phoneNumber:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
...
    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        adminId:
          type: integer
        totalMembers:
          type: integer
        totalGroups:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - adminId
        - totalMembers
        - totalGroups
        - createdAt
    OrganizationDetails:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        adminId:
          type: integer
        totalMembers:
          type: integer
        totalGroups:
          type: integer
        totalContributed:
          type: number
        totalPaidOut:
          type: number
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - adminId
        - totalMembers
        - totalGroups
        - totalContributed
        - totalPaidOut
        - groups
        - createdAt
    CreateOrganizationBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    # GROUP
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        adminId:
          type: integer
        organizationId:
          type: integer
          nullable: true
        contributionAmount:
          type: number
        currency:
          type: string
          default: USD
        schedule:
          type: string
          enum: [weekly, bi-weekly, monthly]
        maxMembers:
          type: integer
        currentCycle:
          type: integer
        currentRotationIndex:
          type: integer
        status:
          type: string
          enum: [active, paused, completed]
        totalMembers:
          type: integer
        paidCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - adminId
        - contributionAmount
        - schedule
        - maxMembers
        - currentCycle
        - currentRotationIndex
        - status
        - totalMembers
        - paidCount
        - createdAt
...
    Contribution:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        groupId:
          type: integer
        cycleId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [paid, pending, failed]
        paidAt:
          type: string
          format: date-time
          nullable: true
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        group:
          $ref: "#/components/schemas/Group"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - groupId
        - cycleId
        - amount
        - status
        - createdAt
    PayContributionBody:
      type: object
      properties:
        groupId:
          type: integer
        cycleId:
          type: integer
      required:
        - groupId
        - cycleId
    ContributionListResponse:
      type: object
      properties:
        contributions:
          type: array
          items:
            $ref: "#/components/schemas/Contribution"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - contributions
        - total
        - page
        - limit

    # PAYOUT
...